+++ setup.cfg
@@ -17,6 +17,8 @@ omit = site-packages
 
 [coverage:report]
 precision = 2
+omit =
+    data_extractor/abc.py
 # Regexes for lines to exclude from consideration
 exclude_lines =
     # Have to re-enable the standard pragma
