+++ tests/utils.py
@@ -1,5 +1,5 @@
 # First Party Library
-from data_extractor.abc import AbstractExtractors
+from data_extractor.core import AbstractExtractors
 
 
 def is_built(obj: AbstractExtractors = None) -> bool:
